+++ packages/redux/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* Switch from `action` to `navigation` (which contains an `action` as well as the previous `response`).
+
 ## 1.0.0-beta.2
 
 * Switch to `router` from `config`.